+++ include/citrius.h
@@ -4,6 +4,7 @@
 #include "nn/linear.h"
 #include "nn/module.h"
 #include "operations.h"
+#include "shape_operations.h"
 #include "tensor.h"
 #include "tensor_factory.h"
 #include "types.h"