+++ include/tensor_factory.h
@@ -6,12 +6,17 @@
 
 namespace citrius {
 
-class TensorFactory {
-public:
-    static Tensor empty(Shape shape, DType dtype = DType::Float32, Device device = Device::cpu());
-    static Tensor from_vector(const std::vector<float>& values, Device device = Device::cpu());
-    static Tensor from_vector(const std::vector<float>& values, Shape shape, Device device = Device::cpu());
-    static Tensor to(const Tensor& tensor, Device device);
-};
+Tensor empty(
+    Shape shape,
+    DType dtype = DType::Float32,
+    Device device = Device::cpu());
+Tensor from_vector(
+    const std::vector<float>& values,
+    Device device = Device::cpu());
+Tensor from_vector(
+    const std::vector<float>& values,
+    Shape shape,
+    Device device = Device::cpu());
+Tensor to(const Tensor& tensor, Device device);
 
 } // namespace citrius